06/01-00-agent-langgraph/requirements.txt
aiofiles==24.1.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
gunicorn
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
langchain-core>=0.1.52 # Ensure compatibility
langchain-openai>=0.1.8 # Ensure compatibility
langgraph>=0.0.57 # Latest as of mid-June 2025
chromadb>=0.5.0 # Updated from 0.4.24
MarkupSafe==3.0.2
priority==2.0.0
python-dotenv
Quart==0.20.0
uvicorn
Werkzeug==3.1.3
wsproto==1.2.0
